Optimize image data type casting
Refactor the MRI image and segmentation data handling across multiple utilities to ensure consistency in data type casting. Now, image data is uniformly cast to 'float64', and segmentation data to 'uint8' immediately after loading, streamlining the data pipeline and improving maintainability. As part of the augmentation processes, the output image's data type now matches the original, preserving the data integrity. This resolves potential issues with data type mismatches, especially when dealing with segmentations and their respective IDs. Additionally, removed commented code and dead lines for clarity.

+++ src/totalsegmri/utils/fix_csf_label.py
@@ -128,10 +128,7 @@ def fix_csf_label(seg_path, segs_path, cord_label, csf_label, output_path, seg_s
 
     # Load segmentation
     seg = nib.load(seg_path)
-    seg_data = seg.get_fdata()
-
-    # Convert data to uint8 to avoid issues with segmentation IDs
-    seg_data = seg_data.astype(np.uint8)
+    seg_data = seg.get_fdata().astype(np.uint8)
 
     # Create an array of x indices
     x_indices = np.broadcast_to(np.arange(seg_data.shape[0])[..., np.newaxis, np.newaxis], seg_data.shape)
